Make sure the vue 2.7 version of vue-demi is used
Due to the way vue-demi works, it doesn't work properly in a nix build. It replaces the files in `vue-demi/lib` with the relevant files for the vue version in a post-install. This doesn't happen in our nix build, and making it happen is more than annoying. By editing the vite config to always resolve to the 2.7 version, we can avoid the nix build erroring out.